通知、任务的延迟消息、重试改为使用rabbitmq_delayed_message插件。

+++ framework-amqp-rabbit/src/main/java/me/insidezhou/southernquiet/amqp/rabbit/DelayedMessage.java
@@ -13,11 +13,11 @@ import java.lang.annotation.*;
 @Documented
 public @interface DelayedMessage {
     @AliasFor("delay")
-    long value() default 0;
+    int value() default 0;
 
     /**
      * 单位：毫秒。
      */
     @AliasFor("value")
-    long delay() default 0;
+    int delay() default 0;
 }
